# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * True Echo VR Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CustomerStatus =
  (typeof CustomerStatus)[keyof typeof CustomerStatus];

export const CustomerStatus = {
  active: "active",
  inactive: "inactive",
  trial: "trial",
} as const;

export interface PointToChild {
  /**
   * @minLength 1
   * @maxLength 80
   */
  label: string;
}

export interface PointToItem {
  /**
   * @minLength 1
   * @maxLength 80
   */
  label: string;
  /** @maxItems 50 */
  children?: PointToChild[];
}

/**
 * @maxItems 50
 */
export type PointToObjects = PointToItem[];

export interface Customer {
  id: string;
  name: string;
  contactEmail: string;
  programVersion: string;
  headsetCount: number;
  activeHeadsets: number;
  status: CustomerStatus;
  pointToObjects: PointToObjects;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
  contactEmail: string;
  programVersion: string;
}

export interface LocationSummary {
  id: string;
  customerId: string;
  name: string;
  createdAt: string;
  qrCodeCount: number;
  lastCalibratedAt?: string;
  lastCalibratedByHeadsetId?: string;
}

export interface CreateLocationBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
}

export interface Vec3 {
  x: number;
  y: number;
  z: number;
}

export interface Vec4 {
  x: number;
  y: number;
  z: number;
  w: number;
}

export interface QrCode {
  id: string;
  locationId: string;
  qrValue: string;
  posX: number;
  posY: number;
  posZ: number;
  rotX: number;
  rotY: number;
  rotZ: number;
  rotW: number;
  calibratedAt: string;
  headsetId?: string;
}

export interface LocationQrData {
  locationId: string;
  locationName: string;
  qrCodes: QrCode[];
  lastCalibratedAt?: string;
  lastCalibratedByHeadsetId?: string;
}

export interface QrCodeImportItem {
  /** @minLength 1 */
  qrValue: string;
  position: Vec3;
  rotation: Vec4;
}

export interface ImportQrCodesBody {
  headsetId?: string;
  qrCodes: QrCodeImportItem[];
}

export interface QrDictionaryEntry {
  id: string;
  customerId: string;
  qrValue: string;
  name: string;
  updatedAt: string;
}

export interface CreateQrDictionaryEntryBody {
  /**
   * @minLength 1
   * @maxLength 100
   */
  qrValue: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
}

export interface HeadsetStartupQrCode {
  qrValue: string;
  name?: string;
  posX: number;
  posY: number;
  posZ: number;
  rotX: number;
  rotY: number;
  rotZ: number;
  rotW: number;
}

export type HeadsetStartupDataNameDictionaryItem = {
  qrValue: string;
  name: string;
};

export interface HeadsetStartupData {
  locationId: string;
  locationName: string;
  qrCodes: HeadsetStartupQrCode[];
  nameDictionary: HeadsetStartupDataNameDictionaryItem[];
}

export type HeadsetStatus = (typeof HeadsetStatus)[keyof typeof HeadsetStatus];

export const HeadsetStatus = {
  online: "online",
  offline: "offline",
  busy: "busy",
} as const;

export interface Headset {
  id: string;
  serialNumber: string;
  label: string;
  customerId: string;
  customerName: string;
  status: HeadsetStatus;
  batteryLevel: number;
  firmwareVersion: string;
  lastSeen: string;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  waiting: "waiting",
  active: "active",
  ended: "ended",
} as const;

export interface Session {
  id: string;
  headsetId: string;
  headsetLabel: string;
  adminRole?: string;
  techRole?: string;
  status: SessionStatus;
  roomCode: string;
  startedAt: string;
  endedAt?: string;
}

export type CreateSessionBodyRole =
  (typeof CreateSessionBodyRole)[keyof typeof CreateSessionBodyRole];

export const CreateSessionBodyRole = {
  admin: "admin",
  tech: "tech",
} as const;

export interface CreateSessionBody {
  headsetId: string;
  role: CreateSessionBodyRole;
}

export type MessageSenderRole =
  (typeof MessageSenderRole)[keyof typeof MessageSenderRole];

export const MessageSenderRole = {
  admin: "admin",
  tech: "tech",
} as const;

export interface Message {
  id: string;
  sessionId: string;
  senderRole: MessageSenderRole;
  content: string;
  sentAt: string;
}

export type SendMessageBodySenderRole =
  (typeof SendMessageBodySenderRole)[keyof typeof SendMessageBodySenderRole];

export const SendMessageBodySenderRole = {
  admin: "admin",
  tech: "tech",
} as const;

export interface SendMessageBody {
  senderRole: SendMessageBodySenderRole;
  content: string;
}

export type DashboardSummaryCustomersByStatus = {
  active: number;
  inactive: number;
  trial: number;
};

export interface DashboardSummary {
  totalCustomers: number;
  totalHeadsets: number;
  activeHeadsets: number;
  activeSessions: number;
  customersByStatus: DashboardSummaryCustomersByStatus;
}

export type DeleteLocation200 = {
  ok: boolean;
};

export type ClearLocationQrCodes200 = {
  ok: boolean;
};

export type DeleteQrDictionaryEntry200 = {
  ok: boolean;
};

export type ListHeadsetsParams = {
  status?: ListHeadsetsStatus;
};

export type ListHeadsetsStatus =
  (typeof ListHeadsetsStatus)[keyof typeof ListHeadsetsStatus];

export const ListHeadsetsStatus = {
  online: "online",
  offline: "offline",
  busy: "busy",
} as const;

export type GetHeadsetStartupDataParams = {
  locationId: string;
};